Moved screen position to position component
